requirements.txt
anyio==3.6.2
cachetools==5.2.0
click==8.1.3
fastapi==0.87.0
h11==0.14.0
idna==3.4
loguru==0.6.0
pydantic==1.10.2
sniffio==1.3.0
typing_extensions==4.4.0
uvicorn==0.20.0
sse-starlette==1.2.1
PyYAML==5.4.1
hyper==0.7.0
hyperframe==3.2.0
PyJWT==2.6.0
cryptography==38.0.4
base58==2.1.1
py-near==1.0.6
pillow==9.5.0